Domains: improve the UX by better enforcement on subdomains
Use text fields with suffixes to better visualize what a valid input is
like. To be on the safe side, properly handle also when the user ignores
the visual queues and inputs the whole domain (test vs
test.example.com). In both cases, the saved name will be
test.example.com.