Framework: Fix loading of the `catch-js-errors` module.
This required a change in how the module entries are built.
Before that, webpack module system was included into the `vendor` module.
Since we want to load `catch-js-errors` first, we need to load the module system before that
so it is now extracted into a `common.js` file.

For the next changes, I think it would be probably better to exclude this file from webpack's build
and simply include it like third party services (olark...) in the page.